+++ server/hazard-service/src/main/java/edu/illinois/ncsa/incore/service/hazard/models/eq/Earthquake.java
@@ -11,7 +11,6 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import dev.morphia.annotations.Id;
-import dev.morphia.annotations.Property;
 import edu.illinois.ncsa.incore.common.data.models.jackson.JsonDateSerializer;
 import org.bson.types.ObjectId;
 
@@ -25,7 +24,6 @@ import java.util.List;
 @XmlRootElement
 public abstract class Earthquake {
     @Id
-    @Property("_id")
     private ObjectId id;
 
     public void setName(String name) {