+++ server/hazard-service/src/main/java/edu/illinois/ncsa/incore/service/hazard/models/tornado/Tornado.java
@@ -12,7 +12,6 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import dev.morphia.annotations.Id;
-import dev.morphia.annotations.Property;
 import edu.illinois.ncsa.incore.common.data.models.jackson.JsonDateSerializer;
 import edu.illinois.ncsa.incore.service.hazard.models.tornado.utils.TornadoCalc;
 import org.bson.json.JsonParseException;
@@ -29,7 +28,6 @@ import java.util.List;
 @XmlRootElement
 public class Tornado {
     @Id
-    @Property("_id")
     private ObjectId id;
     // Friendly name of Tornado hazard
     private String name;