+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/dao/MongoProjectDBRepository.java
@@ -16,8 +16,8 @@ import dev.morphia.Morphia;
 import dev.morphia.mapping.DiscriminatorFunction;
 import dev.morphia.mapping.MapperOptions;
 import dev.morphia.query.Query;
-import dev.morphia.query.experimental.filters.Filters;
-import dev.morphia.query.experimental.updates.UpdateOperators;
+import dev.morphia.query.filters.Filters;
+import dev.morphia.query.updates.UpdateOperators;
 import edu.illinois.ncsa.incore.service.project.models.Project;
 import org.bson.types.ObjectId;