+++ server/semantics-service/src/main/java/edu/illinois/ncsa/incore/service/semantics/model/Type.java
@@ -14,7 +14,6 @@ import java.util.stream.Collectors;
 @Entity("Type")
 public class Type {
     @Id
-    @Property("_id")
     private ObjectId id;
 
     @Property("@context")