+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/controllers/ToolController.java
@@ -42,7 +42,7 @@ import java.util.stream.Collectors;
 @OpenAPIDefinition(
     info = @Info(
         description = "IN-CORE Project Service for creating and accessing tools",
-        version = "1.29.0",
+        version = "1.30.0",
         title = "IN-CORE v2 Project Service API",
         contact = @Contact(
             name = "IN-CORE Dev Team",